Move code to the src folder
