--- lib/index.js
@@ -1,6 +1,0 @@
-/**
- * @fileoverview Helps in writing code with css-modules
- * @author Dmitry Zakharov
- */
-
-export { rules } from './rules/index';
