+++ package.json
@@ -13,14 +13,15 @@
   "scripts": {
     "dev": "rollup --config rollup.config.js --watch",
     "build": "rollup --config rollup.config.js",
-    "test": "mocha tests --recursive",
+    "test": "echo FIXME:",
     "lint": "ESLINT_CLI=true eslint --cache",
-    "lint:fix": "npm run lint -- --fix --ext .js,.jsx .",
-    "test-staged": "lint-staged"
+    "lint:fix": "npm run lint -- --fix --ext .js,.jsx ./src",
+    "prettify": "prettier --write",
+    "pre-commit": "lint-staged && npm run test"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged && npm run test"
+      "pre-commit": "npm run pre-commit"
     }
   },
   "lint-staged": {
