+++ src/rules/consistentImportName/index.js
@@ -1,0 +1,1 @@
+export { consistentImportNameRule } from './consistentImportName.rule';
